Refactor short track helpers
